Add a module with many logos
Added image files for many NYU logos, for the university and for several
schools, black, white, and color, short, long, and stacked.  It's not
the complete list but it's a start.

There is a logos.tex file to showcase the logos.  At this point it
doesn't showcase all the logos in the modules, either.

The build.lua script will install the logos and documentation.